Move MetadataRecord class to MetadataRecord.cs
